// src/modules/auth/enums/user-role.enum.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  USER = 'USER',
  MODERATOR = 'MODERATOR',
  LAND_REGISTRAR = 'LAND_REGISTRAR',
  NOTARY = 'NOTARY',
  ANALYST = 'ANALYST',
  GOVERNMENT_OFFICIAL = 'GOVERNMENT_OFFICIAL',
  DEVELOPER = 'DEVELOPER',
  AUDITOR = 'AUDITOR',
  GUEST = 'GUEST',
  RLMUA_OFFICIAL = "RLMUA_OFFICIAL",
  COMPLIANCE_OFFICER = 'COMPLIANCE_OFFICER',
  SYSTEM = 'SYSTEM',
  LAND_OFFICER = 'LAND_OFFICER',
  DAO_MEMBER = 'DAO_MEMBER',
  DAO_EXECUTOR = 'DAO_EXECUTOR',
  PROPOSER = 'PROPOSER',
  ORACLE_OPERATOR = 'ORACLE_OPERATOR'
}